refactor(phase-base): move BasePhase class to Phases/Base
Phase 12b sub-step 4/4: relocate the abstract `BasePhase` class
from `Pipeline/Types/BasePhase.ts` (where it lived only because the
first-cut refactor needed a parking spot) to its semantically correct
home next to its module-private collaborators at
`Pipeline/Phases/Base/BasePhase.ts`.

Layout after this commit (the four-file Phases/Base cluster):

  * BasePhase.ts             -- the abstract Template Method class
                                (run + runPre/runAction/runPost/runFinal
                                + takePhaseScreenshot) and the public
                                `IsPrePayloadValid` brand.
  * BootstrapContextBuilder  -- (from Commit 3)
  * ActionContextBuilder     -- (from Commit 3)
  * BalanceContextSlice      -- (from Commit 3)
  * HandoffHelpers           -- (from Commit 3)
  * PhaseTrace               -- (from Commit 3)

The legacy `Types/BasePhase.ts` shrinks to a 5-line @deprecated shim
that re-exports `default`, `BasePhase`, and `type IsPrePayloadValid`
from the new path so every existing importer keeps working without
source changes through the v8.5 release window. The shim is annotated
for removal in v8.6 alongside the matching `Mediator/Timing/TimingConfig`
barrel cleanup.

Import-path remap inside the moved class file:
  * `../../Base/...`     -> `../../../Base/...`
  * `../Mediator/...`    -> `../../Mediator/...`
  * `../Phases/Base/...` -> `./...`
  * `./<TypesSibling>`   -> `../../Types/<TypesSibling>`

Validation:
  * tsc --noEmit clean.
  * npm run lint:architecture src/Scrapers/Pipeline -- exits 0.
  * Targeted jest run
    (BasePhase|Pipeline/Types|Phases/Init/InitBrowserSetup|Architecture)
    -- 345/345 tests pass across 21 suites.
  * tsup rebuild produced lib/index.d.ts SHA-256 byte-identical to
    pre-commit-3 baseline (221E1611...AB055, 14823 bytes) -- public
    surface preserved.
