+++ src/Scrapers/Pipeline/Types/BasePhase.ts
@@ -1,422 +1,23 @@
 /**
- * Abstract BasePhase — Template Method for the 4-stage phase protocol.
- * PRE -> ACTION -> POST -> FINAL. Each stage returns Procedure<IPipelineContext>.
+ * Legacy import path for the abstract `BasePhase` class.
  *
- * run() is the ONLY entry point — bakes in Guard Clauses (Rule #15).
- * ACTION receives IActionContext (sealed — no discovery methods).
- * TypeScript compiler refuses resolveField/resolveVisible in action().
+ * <p>Phase 12b sub-step 4/4 (2026-06) moved the class body to its
+ * semantically correct home at {@link "../Phases/Base/BasePhase.js"}
+ * alongside its module-private helpers. This file remains as a
+ * back-compat shim for the v8.5 release window so existing callers of
+ * the historic `Pipeline/Types/BasePhase.js` path keep working without
+ * source changes. All three previously-public symbols are re-exported
+ * verbatim:
  *
- * <p>Phase 12b sub-step 3/4 (2026-06): the pure helpers that used to
- * live in this file now sit beside their semantically correct phase
- * neighbours under `Pipeline/Phases/Base/`. The class body still lives
- * here for one more sub-step so existing imports of `BasePhase` /
- * `IsPrePayloadValid` from this path keep working unchanged; the
- * class itself moves in Commit 4. See the per-helper file headers for
- * the rationale behind each split.
+ *   * `default` (the BasePhase class as the default export)
+ *   * `BasePhase` (the same class as a named export)
+ *   * `IsPrePayloadValid` (the Brand<boolean> PRE-payload validity tag)
+ *
+ * No runtime cost: barrel re-exports are erased by tsup at bundle
+ * time, so `lib/index.{cjs,mjs,d.ts}` byte-content is unchanged.
+ *
+ * @deprecated since v8.5 — import from `../Phases/Base/BasePhase.js`
+ *   directly. This shim will be removed in v8.6 once the cluster
+ *   consolidation finishes.
  */
-
-import { ScraperErrorTypes } from '../../Base/ErrorTypes.js';
-import { safeScreenshot } from '../Mediator/Browser/SafeScreenshot.js';
-import type { IPreludeSpec } from '../Mediator/Elements/PagePrelude.js';
-import { awaitPagePrelude, PRELUDE_NONE } from '../Mediator/Elements/PagePrelude.js';
-import { buildActionContext } from '../Phases/Base/ActionContextBuilder.js';
-import { logHandoffSummary } from '../Phases/Base/HandoffHelpers.js';
-import { PHASE_STAGE_EVENT, RESULT_TAG, traceTag } from '../Phases/Base/PhaseTrace.js';
-import { setActivePhase, setActiveStage } from './ActiveState.js';
-import type { Brand } from './Brand.js';
-import { isMockTimingActive } from './Debug.js';
-import { dumpFixtureHtml } from './FixtureCapture.js';
-import { mockPolicyFor } from './MockPhasePolicy.js';
-import type { PhaseName } from './Phase.js';
-import type { IActionContext, IPipelineContext } from './PipelineContext.js';
-import type { Procedure } from './Procedure.js';
-import { fail, succeed } from './Procedure.js';
-import { screenshotPath } from './RunLabel.js';
-
-/** PRE-payload validation outcome — branded for Rule #15. */
-export type IsPrePayloadValid = Brand<boolean, 'IsPrePayloadValid'>;
-
-/** Abstract base for all pipeline phases. */
-abstract class BasePhase {
-  /** Phase identifier — must match the pipeline execution order. */
-  public abstract readonly name: PhaseName;
-
-  /** Per-stage prelude table — subclasses override `prelude` to opt in. */
-  private readonly _basePreludeSpecs: Record<'PRE' | 'ACTION' | 'POST' | 'FINAL', IPreludeSpec> = {
-    PRE: PRELUDE_NONE,
-    ACTION: PRELUDE_NONE,
-    POST: PRELUDE_NONE,
-    FINAL: PRELUDE_NONE,
-  };
-
-  /**
-   * ACTION — the core execution. Receives SEALED context (no discovery).
-   * Subclasses MUST implement. Compiler rejects resolveField/resolveVisible.
-   * Returns Procedure<IActionContext> — runAction merges back into IPipelineContext.
-   * @param ctx - Sealed action context from buildActionContext.
-   * @param input - Same as ctx.
-   * @returns Updated action context or failure.
-   */
-  public abstract action(
-    ctx: IActionContext,
-    input: IActionContext,
-  ): Promise<Procedure<IActionContext>>;
-
-  /**
-   * Template Method — the ONLY way to execute a phase.
-   * Enforces PRE -> ACTION -> POST -> FINAL with Guard Clauses.
-   * ACTION receives sealed IActionContext (no discovery).
-   * Bookended by automatic phase-level diagnostic screenshots
-   * (`<bank>-<phase>-pre-<ts>.png` before PRE, `<bank>-<phase>-post-<ts>.png`
-   * after FINAL). Both no-op outside trace mode (gated by RunLabel).
-   * @param ctx - Pipeline context at phase entry.
-   * @returns Final context after all 4 stages, or first failure.
-   */
-  public run(ctx: IPipelineContext): Promise<Procedure<IPipelineContext>> {
-    setActivePhase(this.name);
-    return this.runStages(ctx, ctx.logger);
-  }
-
-  /**
-   * PRE — discovery step. Full mediator access.
-   * Default: pass through unchanged. Tagged with the active phase name so
-   * subclasses inherit a real `this`-using body and `class-methods-use-this`
-   * is satisfied without the legacy `void this.name` workaround.
-   * @param _ctx - Pipeline context.
-   * @param input - Pipeline context to pass through.
-   * @returns Succeed with input (no-op default).
-   */
-  public pre(
-    _ctx: IPipelineContext,
-    input: IPipelineContext,
-  ): Promise<Procedure<IPipelineContext>> {
-    return this.passThrough(input);
-  }
-
-  /**
-   * POST — validation after action. Full context restored.
-   * @param _ctx - Pipeline context.
-   * @param input - Pipeline context to pass through.
-   * @returns Succeed with input (no-op default).
-   */
-  public post(
-    _ctx: IPipelineContext,
-    input: IPipelineContext,
-  ): Promise<Procedure<IPipelineContext>> {
-    return this.passThrough(input);
-  }
-
-  /**
-   * FINAL — readiness signal. Full context.
-   * @param _ctx - Pipeline context.
-   * @param input - Pipeline context to pass through.
-   * @returns Succeed with input (no-op default).
-   */
-  public final(
-    _ctx: IPipelineContext,
-    input: IPipelineContext,
-  ): Promise<Procedure<IPipelineContext>> {
-    return this.passThrough(input);
-  }
-
-  /**
-   * Phase-name accessor — subclasses call this from no-this overrides
-   * (`pre`/`action`/`post`/`final`) to satisfy `class-methods-use-this`
-   * without resorting to the `void this.name` workaround that S3735 flags.
-   * @returns Phase name.
-   */
-  protected phaseName(): PhaseName {
-    return this.name;
-  }
-
-  /**
-   * Pass-through helper used by the PRE/POST/FINAL defaults. Tags the
-   * payload with the active phase name so the inherited override
-   * implicitly references `this`, keeping `class-methods-use-this`
-   * happy without `void this.name`.
-   * @param input - Pipeline context to forward unchanged.
-   * @returns Succeed with input wrapped in a resolved promise.
-   */
-  protected passThrough(input: IPipelineContext): Promise<Procedure<IPipelineContext>> {
-    input.logger.debug({ message: `[${this.name}] pass-through` });
-    const result = succeed(input);
-    return Promise.resolve(result);
-  }
-
-  /**
-   * Validate PRE produced a valid discovery payload for ACTION.
-   * Override per phase. Default: no validation (INIT, TERMINATE).
-   * @param ctx - Context after PRE completed.
-   * @returns True if payload valid for ACTION.
-   */
-  protected validatePrePayload(ctx: IPipelineContext): IsPrePayloadValid {
-    return (Boolean(ctx) && this.name.length > 0) as IsPrePayloadValid;
-  }
-
-  /**
-   * Declarative page-readiness opt-in per stage. Override to request a
-   * `'dom'` or `'spa'` prelude wait before this phase's PRE / ACTION /
-   * POST / FINAL handler executes. Default {@link PRELUDE_NONE} keeps
-   * phases that do not navigate at zero overhead. The
-   * {@link "../Mediator/Elements/PagePrelude.js"} `awaitPagePrelude`
-   * helper fires the wait + emits structured telemetry — phases do
-   * not call it directly.
-   *
-   * @param stage - The stage about to execute (uppercase enum tag).
-   * @returns Prelude specification for this stage; default `PRELUDE_NONE`.
-   */
-  protected prelude(stage: 'PRE' | 'ACTION' | 'POST' | 'FINAL'): IPreludeSpec {
-    return this._basePreludeSpecs[stage];
-  }
-
-  /**
-   * Drive the 4-stage protocol — split out so run() can bookend
-   * screenshots without losing readability.
-   * @param ctx - Pipeline context at phase entry.
-   * @param log - Logger instance.
-   * @returns Final context after all 4 stages, or first failure.
-   */
-  private async runStages(
-    ctx: IPipelineContext,
-    log: IPipelineContext['logger'],
-  ): Promise<Procedure<IPipelineContext>> {
-    setActiveStage('PRE');
-    const pre = await this.runPre(ctx, log);
-    if (!pre.success) {
-      await this.takePhaseScreenshot(ctx, 'pre-fail');
-      return pre;
-    }
-    if (!this.validatePrePayload(pre.value)) return this.contractViolation();
-    await this.takePhaseScreenshot(pre.value, 'pre-done');
-    return this.runStagesAfterPre(ctx, pre.value, log);
-  }
-
-  /**
-   * Drive ACTION → POST → FINAL with a screenshot after each stage success.
-   * Split out so runStages stays inside the 10-line method ceiling.
-   * @param ctx - Original phase-entry context (for stages that need it).
-   * @param input - Context produced by PRE (validated payload).
-   * @param log - Logger instance.
-   * @returns Final phase context, or first stage failure.
-   */
-  private async runStagesAfterPre(
-    ctx: IPipelineContext,
-    input: IPipelineContext,
-    log: IPipelineContext['logger'],
-  ): Promise<Procedure<IPipelineContext>> {
-    const action = await this.runAction(input, log);
-    if (!action.success) return this.snapshotAndReturn(input, action, 'action-fail');
-    await this.takePhaseScreenshot(action.value, 'action-done');
-    const post = await this.runPost(ctx, action.value, log);
-    if (!post.success) return this.snapshotAndReturn(action.value, post, 'post-fail');
-    await this.takePhaseScreenshot(post.value, 'post-done');
-    const finalResult = await this.runFinal(ctx, post.value, log);
-    if (!finalResult.success) return this.snapshotAndReturn(post.value, finalResult, 'final-fail');
-    await this.takePhaseScreenshot(finalResult.value, 'final-done');
-    return finalResult;
-  }
-
-  /**
-   * Capture a diagnostic screenshot AND fixture HTML dump for this phase.
-   * No-op when (a) not in trace mode (`screenshotPath` returns empty), or
-   * (b) the phase has no browser attached (INIT before launch / TERMINATE
-   * after teardown / headless api-direct phases). The fixture HTML dump
-   * is gated separately by DUMP_FIXTURES_DIR — see FixtureCapture.ts.
-   * Called automatically by `runStages` after each successful stage —
-   * phases never invoke this directly. Four bookend points per phase,
-   * one per stage output: 'pre-done' (after PRE), 'action-done' (after
-   * ACTION), 'post-done' (after POST), 'final-done' (after FINAL — phase
-   * exit, the state next phase's PRE will see).
-   * @param ctx - Pipeline context at the bookend.
-   * @param suffix - Stage-output marker: 'pre-done' / 'action-done' /
-   *   'post-done' / 'final-done'.
-   * @returns True when a screenshot was didCapture, false on no-op skip
-   * (no browser attached, or off-trace path resolution returned empty).
-   */
-  private async takePhaseScreenshot(
-    ctx: IPipelineContext,
-    suffix:
-      | 'pre-done'
-      | 'action-done'
-      | 'post-done'
-      | 'final-done'
-      | 'pre-fail'
-      | 'action-fail'
-      | 'post-fail'
-      | 'final-fail',
-  ): Promise<boolean> {
-    if (!ctx.browser.has) return false;
-    const label = `${this.name}-${suffix}`;
-    const target = screenshotPath(ctx.companyId, label);
-    if (!target) return false;
-    const page = ctx.browser.value.page;
-    const didCapture = await safeScreenshot(page, { path: target, fullPage: false });
-    if (didCapture) ctx.logger.debug({ message: `screenshot: ${target}` });
-    await dumpFixtureHtml(ctx, label);
-    return didCapture;
-  }
-
-  /**
-   * Bundle a failure-snapshot capture with the early-return so each
-   * failing-stage branch in {@link runStagesAfterPre} stays inside the
-   * 10-line method ceiling. Forensic-grade only — the screenshot is
-   * non-fatal: on disk-write failure the original Procedure result is
-   * still returned unchanged.
-   *
-   * @param ctx - Pipeline context to screenshot (the LAST successful
-   *   stage's output; {@link takePhaseScreenshot} no-ops without browser).
-   * @param result - Original failure Procedure to forward to the caller.
-   * @param label - Failure-suffix tag (e.g. `'action-fail'`).
-   * @returns The same failure Procedure passed in, after the
-   *   screenshot attempt completes.
-   */
-  private async snapshotAndReturn(
-    ctx: IPipelineContext,
-    result: Procedure<IPipelineContext>,
-    label: 'action-fail' | 'post-fail' | 'final-fail',
-  ): Promise<Procedure<IPipelineContext>> {
-    await this.takePhaseScreenshot(ctx, label);
-    return result;
-  }
-
-  /**
-   * Build contract violation failure for invalid PRE payload.
-   * @returns Failure Procedure with contract message.
-   */
-  private contractViolation(): Procedure<IPipelineContext> {
-    const msg = `STAGE_CONTRACT_VIOLATION: ${this.name}.PRE OK but no target payload`;
-    return fail(ScraperErrorTypes.Generic, msg);
-  }
-
-  /**
-   * Execute PRE stage with trace logging.
-   * MOCK_MODE: consults MockPhasePolicy; short-circuits when policy.pre=true.
-   * @param ctx - Pipeline context.
-   * @param log - Logger instance.
-   * @returns PRE result.
-   */
-  private async runPre(
-    ctx: IPipelineContext,
-    log: IPipelineContext['logger'],
-  ): Promise<Procedure<IPipelineContext>> {
-    setActiveStage('PRE');
-    if (isMockTimingActive() && mockPolicyFor(this.name).pre) {
-      const mocked = succeed(ctx);
-      log.debug({ event: PHASE_STAGE_EVENT, phase: this.name, stage: 'PRE', result: 'OK' });
-      return mocked;
-    }
-    const preSpec = this.prelude('PRE');
-    await awaitPagePrelude(ctx, preSpec);
-    const result = await this.pre(ctx, ctx);
-    log.debug({
-      event: PHASE_STAGE_EVENT,
-      phase: this.name,
-      stage: 'PRE',
-      result: traceTag(result),
-    });
-    return result;
-  }
-
-  /**
-   * Execute ACTION stage with sealed context and trace logging.
-   * action() returns Procedure<IActionContext> — merge back into preVal for POST.
-   * On failure: propagate with fail() (IActionContext failure has same shape).
-   * On success: spread preVal + result.value → full IPipelineContext restored.
-   * @param _ctx - Original pipeline context (unused — stages use preVal/restored/postVal).
-   * @param preVal - Context after PRE.
-   * @param log - Logger instance.
-   * @returns ACTION result merged with full PRE context.
-   */
-  /**
-   * Execute ACTION stage with sealed context and trace logging.
-   * action() returns Procedure<IActionContext> — merge back into preVal.
-   * @param preVal - Context after PRE.
-   * @param log - Logger instance.
-   * @returns ACTION result merged with full PRE context.
-   */
-  private async runAction(
-    preVal: IPipelineContext,
-    log: IPipelineContext['logger'],
-  ): Promise<Procedure<IPipelineContext>> {
-    logHandoffSummary(this.name, preVal, log);
-    setActiveStage('ACTION');
-    if (isMockTimingActive() && mockPolicyFor(this.name).action) {
-      log.debug({ event: PHASE_STAGE_EVENT, phase: this.name, stage: 'ACTION', result: 'OK' });
-      return succeed(preVal);
-    }
-    const actionSpec = this.prelude('ACTION');
-    await awaitPagePrelude(preVal, actionSpec);
-    const actionCtx = buildActionContext(preVal);
-    const result = await this.action(actionCtx, actionCtx);
-    log.debug({
-      event: PHASE_STAGE_EVENT,
-      phase: this.name,
-      stage: 'ACTION',
-      result: RESULT_TAG[String(result.success)],
-    });
-    if (!result.success) return fail(result.errorType, result.errorMessage);
-    const restored: IPipelineContext = { ...preVal, ...result.value };
-    return succeed(restored);
-  }
-
-  /**
-   * Execute POST stage with trace logging.
-   * @param _ctx - Original pipeline context (unused — stages use preVal/restored/postVal).
-   * @param restored - Full context restored after action.
-   * @param log - Logger instance.
-   * @returns POST result.
-   */
-  private async runPost(
-    _ctx: IPipelineContext,
-    restored: IPipelineContext,
-    log: IPipelineContext['logger'],
-  ): Promise<Procedure<IPipelineContext>> {
-    setActiveStage('POST');
-    if (isMockTimingActive() && mockPolicyFor(this.name).post) {
-      log.debug({ event: PHASE_STAGE_EVENT, phase: this.name, stage: 'POST', result: 'OK' });
-      return succeed(restored);
-    }
-    const postSpec = this.prelude('POST');
-    await awaitPagePrelude(restored, postSpec);
-    const result = await this.post(restored, restored);
-    log.debug({
-      event: PHASE_STAGE_EVENT,
-      phase: this.name,
-      stage: 'POST',
-      result: traceTag(result),
-    });
-    return result;
-  }
-
-  /**
-   * Execute FINAL stage with trace logging.
-   * @param _ctx - Original pipeline context (unused — stages use preVal/restored/postVal).
-   * @param postVal - Context after POST.
-   * @param log - Logger instance.
-   * @returns FINAL result.
-   */
-  private async runFinal(
-    _ctx: IPipelineContext,
-    postVal: IPipelineContext,
-    log: IPipelineContext['logger'],
-  ): Promise<Procedure<IPipelineContext>> {
-    setActiveStage('FINAL');
-    if (isMockTimingActive() && mockPolicyFor(this.name).final) {
-      log.debug({ event: PHASE_STAGE_EVENT, phase: this.name, stage: 'FINAL', result: 'OK' });
-      return succeed(postVal);
-    }
-    const finalSpec = this.prelude('FINAL');
-    await awaitPagePrelude(postVal, finalSpec);
-    const result = await this.final(postVal, postVal);
-    log.debug({
-      event: PHASE_STAGE_EVENT,
-      phase: this.name,
-      stage: 'FINAL',
-      result: traceTag(result),
-    });
-    return result;
-  }
-}
-
-export default BasePhase;
-export { BasePhase };
+export { BasePhase, default, type IsPrePayloadValid } from '../Phases/Base/BasePhase.js';
